Added schema file, improved database management

+++ notesApp.py
@@ -1,44 +1,53 @@
-from flask import Flask, render_template, request, session
+from flask import Flask, render_template, request, g
 import sqlite3 as sql
 
 app = Flask(__name__)
-
+DATABASE = 'database.db'
 occurrences = []
+
+def init_db():
+    with app.app_context():
+        db = get_db()
+        with app.open_resource('schema.sql', mode='r') as f:
+            db.cursor().executescript(f.read())
+        db.commit()
+
+def get_db():
+    db = getattr(g, '_database', None)
+    if db is None:
+        db = g._database = sql.connect(DATABASE)
+    return db
+
+
+@app.teardown_appcontext
+def close_connection(exception):
+    db = getattr(g, '_database', None)
+    if db is not None:
+        db.close()
+
+
 @app.route('/')
 def home():
-    try:
-        conn = sql.connect('database.db')
-        print("Opened database successfully")
-
-        conn.execute('CREATE TABLE IF NOT EXISTS notes (id INTEGER PRIMARY KEY AUTOINCREMENT,title TEXT NOT NULL, '
-                     'added TEXT NOT NULL)')
-        conn.execute('CREATE TABLE IF NOT EXISTS items (id INTEGER PRIMARY KEY AUTOINCREMENT, note_id INTEGER NOT '
-                     'NULL,title TEXT NOT NULL, added TEXT NOT NULL , modified TEXT NOT NULL, content TEXT NOT NULL, '
-                     'deleted INT NOT NULL, edited INT NOT NULL, FOREIGN KEY (note_id) REFERENCES notes(id),'
-                     'UNIQUE(note_id,title,edited,deleted) ON CONFLICT IGNORE)')
-        print("Tables opened successfully")
-    except sql.Error as error:
-        print("Failed to create table ", error)
-
-    conn.close()
+    init_db()
     return render_template('index.html')
 
 
-@app.route('/addtodb', methods=['POST', 'GET'])
+@app.route('/addtodb', methods=['POST'])
 def addtodb():
     if request.method == 'POST':
         try:
-            conn = sql.connect('database.db')
+            db = get_db()
+            # conn = sql.connect('database.db')
             title = request.form['title']
             content = request.form['content']
             created = request.form['added']
-            conn.execute("INSERT INTO notes (title, added) VALUES(?, ?)", (title, created))
+            db.execute("INSERT INTO notes (title, added) VALUES(?, ?)", (title, created))
 
-            conn.execute(
+            db.execute(
                 "INSERT INTO items (deleted, edited, title,content, added, modified, note_id) SELECT ?,?,?,?,?,?, "
                 "id FROM notes WHERE id=id AND NOT EXISTS (SELECT 1 FROM items WHERE (items.title = notes.title AND "
-                "items.note_id=notes.id))",(0, 0, title, content, created, created))
-            conn.commit()
+                "items.note_id=notes.id))", (0, 0, title, content, created, created))
+            db.commit()
             print("Note added")
 
         except sql.Error as error:
@@ -51,9 +60,8 @@ def addtodb():
 @app.route('/results')
 def results():
     try:
-        conn = sql.connect('database.db')
-        cur = conn.cursor()
-        conn.row_factory = sql.Row
+        db = get_db()
+        cur = db.cursor()
         cur.execute("SELECT * FROM notes")
         data = cur.fetchall()
     except sql.Error as error:
@@ -65,12 +73,12 @@ def results():
 def update():
     if request.method == 'POST':
         try:
-            conn = sql.connect('database.db')
-            cur = conn.cursor()
+            db = get_db()
+            cur = db.cursor()
             id_up = request.form['id']
             title_up = request.form['title']
             cur.execute("SELECT content FROM items WHERE (note_id=? AND title=?)", (id_up, title_up))
-            content_up= cur.fetchall()
+            content_up = cur.fetchall()
             content = content_up.pop()
             content = content[0]
 
@@ -79,23 +87,23 @@ def update():
     return render_template('update.html', id=id_up, title=title_up, content=content)
 
 
-@app.route('/edit', methods=['POST', 'GET'])
+@app.route('/edit', methods=['POST'])
 def edit():
     if request.method == 'POST':
         try:
-            conn = sql.connect('database.db')
+            db = get_db()
             title = request.form['titleN']
             id_up = request.form['idN']
             content = request.form['contentN']
             modified = request.form['modifiedN']
             print(title, content, id_up, modified)
-            cur = conn.cursor()
+            cur = db.cursor()
             occurrences.append(id_up)
             ver = occurrences.count(id_up)
-            conn.execute("INSERT INTO items (note_id, edited, deleted,title,content, added, modified) SELECT note_id, "
+            db.execute("INSERT INTO items (note_id, edited, deleted,title,content, added, modified) SELECT note_id, "
                          "? ,deleted,?,?, added, ? FROM items WHERE note_id=?", (ver, title, content, modified, id_up))
-            conn.execute("UPDATE notes SET title=? WHERE id=? ", (title, id_up))
-            conn.commit()
+            db.execute("UPDATE notes SET title=? WHERE id=? ", (title, id_up))
+            db.commit()
             print("note changed")
             cur.close()
         except sql.Error as error:
@@ -103,51 +111,54 @@ def edit():
     return render_template('added.html')
 
 
-@app.route('/delete', methods=['POST', 'GET'])
+@app.route('/delete', methods=['POST'])
 def delete():
-    try:
-        conn = sql.connect('database.db')
-        cur = conn.cursor()
-        id_del = request.form['delete']
-        cur.execute("DELETE FROM notes WHERE id=?", (id_del,))
-        cur.execute("UPDATE items SET deleted =? WHERE note_id=? ", (1, id_del))
-        conn.commit()
-        data_aft = cur.fetchall()
-    except sql.Error as error:
-        print("Failed to delete record ", error)
+    if request.method == 'POST':
+        try:
+            db = get_db()
+            cur = db.cursor()
+            id_del = request.form['delete']
+            cur.execute("DELETE FROM notes WHERE id=?", (id_del,))
+            cur.execute("UPDATE items SET deleted =? WHERE note_id=? ", (1, id_del))
+            db.commit()
+            data_aft = cur.fetchall()
+        except sql.Error as error:
+            print("Failed to delete record ", error)
     return render_template('delete.html', data=data_aft)
 
 
-@app.route('/details', methods=['POST', 'GET'])
+@app.route('/details', methods=['POST'])
 def details():
-    try:
-        conn = sql.connect('database.db')
-        cur = conn.cursor()
-        id_show = request.form['details']
-        print(id_show)
-        cur.execute("SELECT * FROM items WHERE note_id=? ORDER BY edited ASC", (id_show,))
-        conn.commit()
-        data_aft = cur.fetchall()
-    except sql.Error as error:
-        print("Failed to query details ", error)
+    if request.method == 'POST':
+        try:
+            db = get_db()
+            cur = db.cursor()
+            id_show = request.form['details']
+            print(id_show)
+            cur.execute("SELECT * FROM items WHERE note_id=? ORDER BY edited ASC", (id_show,))
+            db.commit()
+            data_aft = cur.fetchall()
+        except sql.Error as error:
+            print("Failed to query details ", error)
     return render_template('details.html', data=data_aft)
 
 
-@app.route('/readonly', methods=['POST', 'GET'])
+@app.route('/readonly', methods=['POST'])
 def readonly():
-    try:
-        conn = sql.connect('database.db')
-        cur = conn.cursor()
-        title_read = request.form['titleRead']
-        id_read = request.form['idRead']
-        cur.execute("SELECT content FROM items WHERE (note_id=? AND title=?)", (id_read, title_read))
-        conn.commit()
-        content_read = cur.fetchall()
-        content = content_read.pop()
-        content = content[0]
-    except sql.Error as error:
-        print("Failed to query details ", error)
-    return render_template('readonly.html', content=content,  title=title_read)
+    if request.method == 'POST':
+        try:
+            db = get_db()
+            cur = db.cursor()
+            title_read = request.form['titleRead']
+            id_read = request.form['idRead']
+            cur.execute("SELECT content FROM items WHERE (note_id=? AND title=?)", (id_read, title_read))
+            db.commit()
+            content_read = cur.fetchall()
+            content = content_read.pop()
+            content = content[0]
+        except sql.Error as error:
+            print("Failed to query details ", error)
+    return render_template('readonly.html', content=content, title=title_read)
 
 
 if __name__ == '__main__':
